Improve readability of ConcurrentMapProjection
This commit improves the readability of the function `applyDelta()` by extracting the level of abstraction of the actual change into an external function allowing the reader to easily understand what each case would do. In case of a further information, the reader of the code can look into the specifics of the function and maintain the level of its own abstraction in that context.